models/yolo_v3/requirements.txt
opencv-python
numpy
pillow
python-socketio[client]
